Human_detection/requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.2.0
appnope==0.1.4
apyori==1.1.2
asteroid-filterbanks==0.4.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
Babel==2.14.0
bangla==0.0.2
beautifulsoup4==4.12.3
bitarray==2.9.2
black==24.3.0
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
ctranslate2==4.1.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.8
dateparser==1.1.8
debugpy==1.8.1
decorator==4.4.2
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
docstring_parser_fork==0.0.5
einops==0.7.0
encodec==0.1.1
exceptiongroup==1.2.0
executing==2.0.1
fairseq==0.12.2
fastapi==0.108.0
faster-whisper==1.0.1
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.13.1
flake8==7.0.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
g2pkk==0.1.2
gast==0.5.4
gdown==5.1.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==4.33.0
gradio_client==0.17.0
grpcio==1.60.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.10.0
hangul-romanize==0.1.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
hydra-core==1.0.7
HyperPyYAML==1.2.2
identify==2.5.35
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
inflect==7.0.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.20.0
isort==5.13.2
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.3.2
jsonlines==1.2.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.0
lightning==2.2.2
lightning-utilities==0.11.2
lium-diarization-editor==0.1.1
llvmlite==0.41.1
lxml==5.1.0
Mako==1.3.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.8.1
nodeenv==1.8.0
num2words==0.5.13
numba==0.58.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.17.3
openai==1.7.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
optuna==3.6.1
orjson==3.9.10
packaging==23.2
pandas==1.5.3
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
pipline==0.0
pipreqs==0.4.13
platformdirs==4.1.0
plotly==5.22.0
pluggy==1.4.0
pooch==1.8.0
portalocker==2.8.2
pre-commit==3.7.0
preshed==3.0.9
primePy==1.3
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydoclint==0.4.1
pydub==0.25.1
pyflakes==3.2.0
pygame==2.5.2
Pygments==2.17.2
pygtrie==2.5.0
pynndescent==0.5.11
pyparsing==3.1.1
pypinyin==0.50.0
pysbd==0.3.4
PySocks==1.7.1
pytest==7.4.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-mpv==1.0.5
python-multipart==0.0.9
pytorch-lightning==2.2.2
pytorch-metric-learning==2.5.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.4.7
sacrebleu==2.4.0
safetensors==0.4.1
scikit-learn==1.4.2
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.0
SQLAlchemy==2.0.29
srsly==2.4.8
stack-data==0.6.3
starlette==0.32.0.post1
SudachiDict-core==20230927
SudachiPy==0.6.8
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
thinc==8.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.3.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.2
torchvision==0.18.1
tornado==6.4
tqdm==4.66.1
trainer==0.0.36
traitlets==5.14.1
transformers==4.36.2
TTS==0.22.0
typer==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.2.50
ultralytics-thop==2.0.0
umap-learn==0.5.5
Unidecode==1.3.8
urllib3==2.1.0
uvicorn==0.25.0
virtualenv==20.25.3
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
whisperx==3.1.3
wrapt==1.14.1
yamllint==1.35.1
yarg==0.1.9
yarl==1.9.4
yt-dlp==2024.5.27
